Optics.TH.Internal.Product

Documentation

data LensRules

Rules to construct lenses for data fields.

Constructors

LensRules

Fields

_simpleLenses :: Bool

_generateSigs :: Bool

_generateClasses :: Bool

_allowIsos :: Bool

_allowUpdates :: Bool

Allow LensTraversal (otherwise GetterFold)

_lazyPatterns :: Bool

_fieldToDef :: FieldNamer

Type Name -> Field Names -> Target Field Name -> Definition Names

_classyLenses :: ClassyNamer

type FieldNamer

Arguments

= Name
Name of the data type that lenses are being generated for.

-> [Name]
Names of all fields (including the field being named) in the data type.

-> Name
Name of the field being named.

-> [DefName]
Name(s) of the lens functions. If empty, no lens is created for that field.

The rule to create function names of lenses for data fields.

Although it's sometimes useful, you won't need the first two arguments most of the time.

data DefName

Name to give to generated field optics.

Constructors

TopName Name
Simple top-level definition name

MethodName Name Name
makeFields-style class name and method name

type ClassyNamer

Arguments

= Name
Name of the data type that lenses are being generated for.

-> Maybe (Name, Name)
Names of the class and the main method it generates, respectively.

The optional rule to create a class and method around a monomorphic data type. If this naming convention is provided, it generates a "classy" lens.

makeFieldOptics :: LensRules -> Name -> DecsQ

Compute the field optics for the type identified by the given type name. Lenses will be computed when possible, Traversals otherwise.

makeFieldOpticsForDec :: LensRules -> Dec -> DecsQ

makeFieldOpticsForDec' :: LensRules -> Dec -> HasFieldClasses [Dec]

makeFieldLabelsWith :: LensRules -> Name -> DecsQ

Build field optics as labels with a custom configuration.

makeFieldLabelsForDec :: LensRules -> Dec -> DecsQ

type HasFieldClasses = StateT (Set Name) Q

Tracks the field class Names that have been created so far. We consult these so that we may avoid creating duplicate classes.