Optics.TH.Internal.Utils

Documentation

appsT :: TypeQ -> [TypeQ] -> TypeQ

Apply arguments to a type constructor

appsE1 :: ExpQ -> [ExpQ] -> ExpQ

Apply arguments to a function

toTupleT :: [TypeQ] -> TypeQ

Construct a tuple type given a list of types.

toTupleE :: [ExpQ] -> ExpQ

Construct a tuple value given a list of expressions.

toTupleP :: [PatQ] -> PatQ

Construct a tuple pattern given a list of patterns.

conAppsT :: Name -> [Type] -> Type

Apply arguments to a type constructor.

bndrName :: TyVarBndr -> Name

Return Name contained in a TyVarBndr.

newNames

Arguments

:: String
base name

-> Int
count

-> Q [Name]

Generate many new names from a given base name.

eqSubst :: Type -> String -> Q (Type, Pred)

addKindVars :: DatatypeInfo -> Type -> Type

Fill in kind variables using info from datatype type parameters.

quantifyType :: [TyVarBndr] -> Cxt -> Type -> Type

Template Haskell wants type variables declared in a forall, so we find all free type variables in a given type and declare them.

quantifyType' :: Set Name -> [TyVarBndr] -> Cxt -> Type -> Type

This function works like quantifyType except that it takes a list of variables to exclude from quantification.

inlinePragma :: Name -> [DecQ]